remove composed, rename elements
