+++ @polaris/components/src/themes/vars.css.ts
