--- @polaris/composed/README.md
@@ -1,1 +1,0 @@
-# @polaris/composed
