--- @polaris/composed/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@polaris/composed",
-  "version": "0.0.0",
-  "main": "dist/polaris-composed.cjs.js",
-  "module": "dist/polaris-composed.esm.js",
-  "license": "MIT",
-  "peerDependencies": {
-    "react": "^16.8 || ^17.0"
-  }
-}
