--- @polaris/elements/README.md
@@ -1,1 +1,0 @@
-# @polaris/elements
