+++ polaris.shopify.com/components/Heading/Heading.tsx
@@ -1,5 +1,5 @@
 import React, {ComponentProps} from 'react';
-import {Box} from '@polaris/elements';
+import {Box} from '@polaris/components';
 
 export const Heading = (props: ComponentProps<typeof Box>) => (
   <Box {...props} />
